+++ include/mysql/impl/deserialize_row.ipp
@@ -50,7 +50,7 @@ inline Error deserialize_text_value_impl(
 	int parsed = decimals ? sscanf(buffer, "%4d:%2u:%2u.%6u", &hours, &minutes, &seconds, &micros) : // sign adds 1 char
 			                sscanf(buffer, "%4d:%2u:%2u", &hours, &minutes, &seconds);
 	if ((decimals && parsed != 4) || (!decimals && parsed != 3)) return Error::protocol_value_error;
-	micros *= std::pow(10, 6 - decimals);
+	micros *= static_cast<unsigned>(std::pow(10, 6 - decimals));
 	bool is_negative = hours < 0;
 	hours = std::abs(hours);
